Move McpIdentity to main package
