+++ mcp/src/main/java/io/airlift/mcp/McpIdentityMapper.java
@@ -1,6 +1,5 @@
 package io.airlift.mcp;
 
-import io.airlift.mcp.model.McpIdentity;
 import jakarta.servlet.http.HttpServletRequest;
 
 public interface McpIdentityMapper
