+++ mcp/src/main/java/io/airlift/mcp/internal/InternalFilter.java
@@ -11,6 +11,8 @@ import io.airlift.log.Logger;
 import io.airlift.mcp.CancellationController;
 import io.airlift.mcp.ErrorHandler;
 import io.airlift.mcp.McpConfig;
+import io.airlift.mcp.McpIdentity;
+import io.airlift.mcp.McpIdentity.Authenticated;
 import io.airlift.mcp.McpIdentityMapper;
 import io.airlift.mcp.McpMetadata;
 import io.airlift.mcp.model.CallToolRequest;
@@ -21,8 +23,6 @@ import io.airlift.mcp.model.InitializeRequest;
 import io.airlift.mcp.model.JsonRpcRequest;
 import io.airlift.mcp.model.JsonRpcResponse;
 import io.airlift.mcp.model.ListRequest;
-import io.airlift.mcp.model.McpIdentity;
-import io.airlift.mcp.model.McpIdentity.Authenticated;
 import io.airlift.mcp.model.Protocol;
 import io.airlift.mcp.model.ReadResourceRequest;
 import io.airlift.mcp.model.SetLevelRequest;
